requirements.txt
# requirements.txt
Flask>=2.0
pandas>=1.3
scikit-learn>=1.0
supabase>=1.0
python-dotenv>=0.19